Standardized naming for company and ch details in pages. Added support in the controller to create url’s for archiving Fixed issue with babel polypill clashing with trade elements 3.01 Fixed sizing issue for left nav Updated labels and order based on designs
