+++ src/javascripts/index.js
@@ -1,6 +1,4 @@
 /* eslint no-new: 0 */
-require('babel-polyfill')
-
 const ClippedList = require('./clippedlist')
 
 new ClippedList(document.getElementById('interactions-list'), 'See all new interactions', 'See less interactions')
