+++ src/sass/_leftnav.scss
@@ -1,6 +1,5 @@
 .left-nav-container {
 
-  @extend %site-width-container;
   @extend %contain-floats;
   clear: both;
