+++ src/sass/_shame.scss
@@ -88,7 +88,7 @@ hr {
 
   p.footnote {
     color: $grey-1;
-    padding-left: 35px;
+    padding-left: 25px;
     padding-bottom: 10px;
     margin: 0;
   }
