+++ src/views/company/add-step-1.html
@@ -12,7 +12,7 @@
     {% endfor %}
   {% endif %}
 
-  {{ trade.errorPanel( errors, labels=companyDetailLabels ) }}
+  {{ trade.errorPanel( errors, labels=companyDetailsLabels ) }}
   <form id="company-add-form" data-step="1" method="post">
     <input type="hidden" name="_csrf_token" value="{{ csrfToken }}" />
 
@@ -57,6 +57,5 @@
       <a href="/" class="button-link button--link">Cancel</a>
     </div>
   </form>
-  <script src="/javascripts/common.js"></script>
   <script src="/javascripts/companyadd.bundle.js"></script>
 {% endblock %}
