+++ src/views/company/details.html
@@ -6,16 +6,24 @@
 
   <h2 class="heading-medium">Company summary</h2>
 
-  {% if chDisplay %}
+  {% if chDetails %}
     <div class="dark-panel">
-      {{ trade.keyvaluetable(chDisplay, labels=chDetailLabels, keyorder=chDetailsDisplayOrder, readonly=true) }}
+      {{ trade.keyvaluetable(chDetails, labels=chDetailsLabels, keyorder=chDetailsDisplayOrder, readonly=true) }}
       <p class="footnote">Powered by data from <strong>Companies House</strong></p>
     </div>
   {% endif %}
 
-  {% if companyDisplay %}
-    {{ trade.keyvaluetable(companyDisplay, stripey=true, labels=companyDetailLabels, keyorder=companyDetailsDisplayOrder, id="cdms-details") }}
+  {% if companyDetails %}
+    {{ trade.keyvaluetable(companyDetails, stripey=true, labels=companyDetailsLabels, keyorder=companyDetailsDisplayOrder, id="cdms-details") }}
   {% endif %}
 
   <p><a href="/company/{{source}}/{{company.id}}/edit" class="button">Edit</a></p>
+
+  {% if accountManagementDisplay %}
+    <div class="section">
+      <h2 class="heading-medium">Account management</h2>
+      {{ trade.keyvaluetable(accountManagementDisplay, stripey=true, labels=accountManagementDisplayLabels) }}
+      <p class="footnote">You do not have perission to edit the Account management section</strong></p>
+    </div>
+  {% endif %}
 {% endblock %}
