+++ src/views/company/edit-nonuk.html
@@ -6,7 +6,7 @@
   <h1 class="page-heading">Create a company</h1>
   <div class="grid-row">
     <div class="column-three-quarters">
-      {{ trade.errorPanel( errors, labels=companyDetailLabels ) }}
+      {{ trade.errorPanel( errors, labels=companyDetailsLabels ) }}
       <div class="section">
         <div class="for-control key-value key-value--vertical">
           <div id="business-type" class="key-value__key form-label-bold">Business type
@@ -17,8 +17,8 @@
       </div>
 
       <div class="section">
-        {% if chDisplay %}
-          {{ trade.keyvaluetable(chDisplay, labels=chDetailLabels, keyorder=chDetailsDisplayOrder, class="table--small") }}
+        {% if chDetails %}
+          {{ trade.keyvaluetable(chDetails, labels=chDetailsLabels, keyorder=chDetailsDisplayOrder, class="table--small") }}
         {% endif %}
       </div>
 
@@ -136,6 +136,5 @@
 
     </div>
   </div>
-  <script src="/javascripts/common.js"></script>
   <script src="/javascripts/companyedit.bundle.js"></script>
 {% endblock %}
