+++ src/views/company/index.html
@@ -12,7 +12,7 @@
 {% else %}
   {% if company.id %}
     <p>
-      <div class="form-label-bold">{{company.id}}</div>
+      <div class="form-label-bold">{{company.ref}}</div>
       <a href="{{archiveUrl}}">Archive company</a>
     </p>
   {% endif %}
