+++ src/views/company/investmentform.html
@@ -80,7 +80,5 @@
     {% endif %}
   </div>
 </form>
-
-<script src="/javascripts/common.js"></script>
 <script src="/javascripts/companyinvestmenttabform.bundle.js"></script>
 {% endblock %}
