+++ src/views/investment/create.html
@@ -143,7 +143,6 @@
     {#</form>#}
   </div>
 
-  <script src="/javascripts/common.js"></script>
   <script src="/javascripts/createinvestment.bundle.js"></script>
