+++ src/views/investment/index.html
@@ -49,7 +49,6 @@ id="investment-details")
 
   <p class="infostrip hidden" id="new_co_nagger">If you can’t find the company you’re looking for, <a href="/company/add-step-1/">create a new company</a>. Unfortunately you will lose your progress.</p>
 
-  <script src="/javascripts/common.js"></script>
   <script src="/javascripts/investment.bundle.js"></script>
 
 {% endblock %}}
