+++ test/mocha.opts
@@ -1,4 +1,3 @@
 --require test/common.js
 --recursive
 --reporter list
---compilers js:babel-core/register
